% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_stuff.R
\name{anova2kable}
\alias{anova2kable}
\title{Functions to produce kable output from anova()}
\usage{
anova2kable(
  a,
  row_labels,
  table_caption = NULL,
  sig_figs = 3,
  alpha = 0.05,
  p_thresh = 1e-04
)
}
\description{
Functions to produce kable output from anova()
}
