% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_letters.R
\name{bayes_multcompletters}
\alias{bayes_multcompletters}
\title{Bayesian multiple comparison letters from posterior summary}
\usage{
bayes_multcompletters(
  post_cons,
  post_means,
  group_col,
  contrast_col = "contrast",
  mean_col = "emmean",
  pvalue_col = "p_MAP",
  contrast_sep = " - ",
  alpha = 0.05,
  decreasing_letters = FALSE
)
}
\arguments{
\item{post_cons}{Data frame of contrasts, usually an object of class describe_posterior}

\item{post_means}{Data frame of means, used for correctly ordering the contrasts, usually an object of class describe_posterior}

\item{group_col}{Name of column, the levels of which are being compared}

\item{contrast_col}{Name of column with the pairwise contrast names, default is \code{"contrast"}}

\item{mean_col}{Name of column with the means in post_means, default is \code{"emmean"}}

\item{pvalue_col}{Name of column with the Bayesian p-value analog, default is \code{"p_MAP"}}

\item{contrast_sep}{Separator of the two levels in the contrast column, default is \code{" - "}}

\item{alpha}{Significance cutoff, defaults to 0.05 for no good reason}

\item{decreasing_letters}{Defaults to FALSE so that "a" is the lowest mean. If TRUE, "a" is the highest mean.}
}
\description{
Uses output from \code{bayestestR::describe_posterior()} run on \code{emmeans} objects. Based on the Bayesian p-value analog.
}
