% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model_onerow.R
\name{check_model_onerow}
\alias{check_model_onerow}
\title{Custom function to plot three check model plots on one line with common title}
\usage{
check_model_onerow(fit, common_title)
}
\description{
Custom function to plot three check model plots on one line with common title
}
