% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emm_heatmap.R
\name{emm_heatmap}
\alias{emm_heatmap}
\title{Make a heat map from contrasts between estimated marginal means}
\usage{
emm_heatmap(
  emms,
  level_plot_order = NULL,
  bayes = FALSE,
  alpha = 0.95,
  back_tran_fn = c,
  fill_scale_fn = ggplot2::scale_fill_viridis_c,
  fill_scale_trans = "identity",
  asterisk_color = "white",
  axis_labels = "group",
  legend_label = "difference"
)
}
\description{
Make a heat map from contrasts between estimated marginal means
}
