% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emm_quantile.R
\name{emm_quantile}
\alias{emm_quantile}
\title{Function to get quantiles from an emm or contrast object}
\usage{
emm_quantile(x, probs = c(0.5, 0.025, 0.975, 0.05, 0.95, 0.17, 0.83))
}
\description{
Function to get quantiles from an emm or contrast object
}
