% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_letters.R
\name{get_comp_letters}
\alias{get_comp_letters}
\title{Additional improved comparison letter function based on tidyverse}
\usage{
get_comp_letters(
  post_means,
  post_cons,
  diff_byvars = c(),
  diff_levelvar,
  x_var,
  pvalue_var,
  contrast_var = "contrast",
  contrast_sep = " - ",
  p_thresh = 0.05,
  decreasing_letters = FALSE
)
}
\description{
Additional improved comparison letter function based on tidyverse
}
