% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_comparisons.R
\name{get_pairwise_differences}
\alias{get_pairwise_differences}
\title{Calculate all pairwise differences between levels of a variable, by group}
\usage{
get_pairwise_differences(
  dt,
  valuevar,
  groupvar,
  byvars,
  op = `-`,
  reverse = FALSE,
  difference_name = "diff"
)
}
\arguments{
\item{dt}{Input data frame.}

\item{valuevar}{Name of value column that is going to be differenced.}

\item{groupvar}{Name of column for which you want the pairwise differences between its levels.}

\item{byvars}{Vector of column names within each combination of which you want the pairwise differences between the levels of the group variable.}

\item{op}{Pairwise operation. Defaults to subtraction. Supply the operator in backticks.}

\item{reverse}{Whether to reverse the order of the pairwise comparisons. Default is FALSE meaning we subtract meanA-meanB, where meanA is the first of the two levels when sorting the levels of groupvar in increasing order.}

\item{difference_name}{Name of the difference column in the output. Default is "diff".}
}
\description{
Updated version that does not have any dependencies. Works well with the output of \code{tidybayes::add_epred_draws}.
}
