% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_diff_p.R
\name{loo_diff_p}
\alias{loo_diff_p}
\title{Get p-values from standardized differences in loo_compare}
\usage{
loo_diff_p(loo_compare)
}
\description{
Get p-values from standardized differences in loo_compare
}
