% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_comparisons.R
\name{mean_class_pred}
\alias{mean_class_pred}
\title{Mean class mode predictions}
\usage{
mean_class_pred(fit, pred_grid, groupvars)
}
\description{
This function will get predictions from the output of \code{tidybayes::add_epred_draws} for a Bayesian cumulative logistic model similar to what we get from \code{emmeans::emmeans(..., mode = 'mean.class')} for a frequentist CLM. Uses tidyverse.
}
