% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_quantile.R
\name{median_and_QCI}
\alias{median_and_QCI}
\title{More general function to get medians and quantile credible intervals from a data table for multiple grouping variables, multiple quantiles, and multiple columns to be summarized}
\usage{
median_and_QCI(
  dt,
  group_vars,
  value_vars = ".value",
  CI_widths = c(0.8, 0.95, 0.99)
)
}
\description{
More general function to get medians and quantile credible intervals from a data table for multiple grouping variables, multiple quantiles, and multiple columns to be summarized
}
