% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_page_pdf.R
\name{multiple_page_pdf}
\alias{multiple_page_pdf}
\title{Write list of ggplots to a PDF with multiple pages}
\usage{
multiple_page_pdf(
  plot_list,
  filename,
  height = 11,
  width = 8.5,
  rows_per_page,
  cols_per_page
)
}
\description{
Write list of ggplots to a PDF with multiple pages
}
