% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_quantile.R
\name{pred_quantile}
\alias{pred_quantile}
\title{Function to get prediction quantiles and bind x and predicted y together.}
\usage{
pred_quantile(
  x_pred,
  y_pred,
  x_names = NULL,
  probs = c(0.5, 0.005, 0.995, 0.025, 0.975, 0.05, 0.95, 0.17, 0.83)
)
}
\description{
It calculates median and 99\%, 95\%, 90\%, and 66\% quantile-based HPD.
}
