% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_quantile.R
\name{pred_quantile_3D}
\alias{pred_quantile_3D}
\title{Function to get predicted quantiles in 3D (for multinomial responses)}
\usage{
pred_quantile_3D(
  x_pred,
  y_pred,
  y_dim_names = c("quantile", "var2", "var3"),
  response_name = "p",
  probs = c(0.5, 0.005, 0.995, 0.025, 0.975, 0.05, 0.95, 0.17, 0.83)
)
}
\description{
y_pred should be an array.
}
