% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_stuff.R
\name{pretty_sci_not}
\alias{pretty_sci_not}
\title{Function to produce "pretty" scientific notation in HTML or LaTeX style}
\usage{
pretty_sci_not(x, sig_figs = 3, log_thresh = 4, style = "html", minimum = NA)
}
\arguments{
\item{x}{Numeric vector.}

\item{sig_figs}{Number of significant figures for the mantissa. Default 3.}

\item{log_thresh}{Any number with a base-10 exponent (absolute value)
less than this number will not be output in scientific notation. Default 4.}

\item{style}{'html' (default) or 'latex'}

\item{minimum}{if provided, any number less than this value will be set to this
value and preceded by ">"}
}
\value{
Character vector.
}
\description{
Function to produce "pretty" scientific notation in HTML or LaTeX style
}
