% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.csv.better.R
\name{read.csv.better}
\alias{read.csv.better}
\title{read.csv wrapper that detects all space characters}
\usage{
read.csv.better(file_path, ...)
}
\description{
A wrapper function for read.csv that uses perl to convert any possible spaces to a single type
and then reads the CSV with any other arguments to read.csv that you want to input.
}
\examples{
\dontrun{
}

}
