% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all_csvs.R
\name{read_all_csvs}
\alias{read_all_csvs}
\title{Read all CSVs in a directory}
\usage{
read_all_csvs(dir_path)
}
\description{
The resulting R objects will have the same name as the CSVs,
with the .csv suffix removed. Uses fread()
}
