% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_letters.R
\name{readjusted_alpha}
\alias{readjusted_alpha}
\title{Utility function to readjust alpha for doing Sidak corrections within each CLD by group}
\usage{
readjusted_alpha(true_alpha, treatments_total, treatments_per_group)
}
\description{
Utility function to readjust alpha for doing Sidak corrections within each CLD by group
}
