% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na_dt.R
\name{replace_na_dt}
\alias{replace_na_dt}
\title{Replace NA in a data table}
\usage{
replace_na_dt(DT, cols = names(DT), replace_with = 0)
}
\description{
Modified from a function written by Matt Dowle
See answer at https://stackoverflow.com/questions/7235657/fastest-way-to-replace-nas-in-a-large-data-table
}
