% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_black.R
\name{theme_black}
\alias{theme_black}
\title{ggplot2 stealth black on white theme}
\usage{
theme_black(base_size = 12, base_family = "")
}
\description{
I believe this was originally created by Jon Lefcheck.
}
