% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_nest_dt.R
\name{unnest_dt}
\alias{unnest_dt}
\title{Unnest a nested data.table}
\usage{
unnest_dt(dt, col, id, id_vars = NULL)
}
\description{
This function was written by TS Barrett in a document available at https://osf.io/f6pxw/download.
}
