% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_nest_dt.R
\name{group_nest_dt}
\alias{group_nest_dt}
\title{Nest a data.table by group}
\usage{
group_nest_dt(dt, ..., .key = "data", group_vars = NULL)
}
\description{
This function was originally written by TS Barrett in a document available at https://osf.io/f6pxw/download.
QDR edited so that it can also accept a character vector (good for programmatic usage of the function)
}
